% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_fun.R
\name{adjust_levels}
\alias{adjust_levels}
\title{Adjust test level for multiple testing}
\usage{
adjust_levels(level, nhyp, type = "Holm")
}
\arguments{
\item{level}{The test level (numeric value between 0 and 1)}

\item{nhyp}{Number of hypotheses that are tested}

\item{type}{Either 'Holm' for the Holm-procedure or 'BenjYek' for the Benjamini Yekutieli procedure.}
}
\value{
A vector containing the adjusted critical values in ascending order.
}
\description{
Adjust test level for multiple testing
}
\examples{
adjust_levels(0.1, 9, type = "Holm")
adjust_levels(0.1, 9, type = "BenjYek")
}
