% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_fun.R
\name{adjust_pvals}
\alias{adjust_pvals}
\title{Holm-correction of p-values for multiple testing}
\usage{
adjust_pvals(pvals)
}
\arguments{
\item{pvals}{Vector of p-values}
}
\value{
Vector of adjusted p-values
}
\description{
Holm-correction of p-values for multiple testing
}
\examples{
adjust_pvals( c(0.001, 0.0002, 0.3, 0.55) )
}
