% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_biv_fun.R
\name{bootstrap_pairs_bivmod}
\alias{bootstrap_pairs_bivmod}
\title{Bootstrap p-values of test-statistic on pairs}
\usage{
bootstrap_pairs_bivmod(
  data,
  temp.cov,
  B = 300,
  H0 = "ED",
  biv_models = c("log", "alog", "hr"),
  varmeth = "chain",
  reg_of_int = NULL,
  pairs = NULL,
  return_boots = FALSE
)
}
\arguments{
\item{data}{A matrix of observations. Each column corresponds to one station.
Columns should be named or numbered to identify subsets.}

\item{temp.cov}{A numeric vector containing values of the temporal covariate. Must be of length \code{nrow(data)}.}

\item{B}{The number of bootstrap replicates.}

\item{H0}{Character code for the null-hypothesis: \code{ED} for the one of equal distribution;
\code{LS} for the one of a local scaling model.}

\item{biv_models}{A vector containing the names of the max-stable models to fit. Must be a subset of
\verb{gauss, brown, powexp, cauchy, whitmat, bessel}.}

\item{varmeth}{Method for estimating the variance-covariance matrix of the stationwise ML estimators, passed to
\code{\link[findpoolreg]{fit_spat_scalegev}}.
Can be either \code{chain} (the default) for an estimator based
on the multivariate chain rule, or \code{basic} for a very simplistic but faster method.}

\item{reg_of_int}{The region of primary interest for the analysis. Can be omitted if \code{pairs} is provided.}

\item{pairs}{A list of pairs on which to perform the test.}

\item{return_boots}{Logical; whether to return bootstrapped test statistics and p values.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{teststat} The value of the test statistic on the observed data
\item \code{p} The asymptotic p value of the test statistic
\item \code{p_boot} The bootstrapped p-value.
\item \code{bootstrap.B} The number of bootstrap repetitions that were performed.
\item \code{Model_Sel} The max-stable process model that was selected.
\item \code{biv_pars} Parameters of the selected max-stable process model.
\item \code{boot_teststat} (if \code{return_boots = TRUE}) Data frame containing the test statistic and corresponding p-value for each bootstrap replicate.
\item \code{par_h0} The parameter estimates under H0.
\item \code{H0} Character code of the null hypothesis.
\item \code{X1}, \code{X2} The labels of pairs on which the hypothesis was tested.
\item \code{adj_p} (if \code{adj_pvals = TRUE}) The Holm-adjusted p-values.
}
}
\description{
Bootstrap p-values of test-statistic on pairs
}
\examples{
\dontrun{
# generate some data
cvrt <- (1:100)/100
coords <- matrix(20*abs(stats::rnorm(4*2)), ncol = 2)
x <- generateData(seed = 2, n = 100, temp.cov = cvrt, d= 4, locations = coords)
bootres <- bootstrap_pairs_bivmod(data = x, temp.cov = cvrt, varmeth = "chain", B = 200, reg_of_int = 3)
bootres
}
}
