% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_fun.R
\name{bootstrap_scalegev}
\alias{bootstrap_scalegev}
\title{Bootstrap p-values for test-statistic}
\usage{
bootstrap_scalegev(
  data,
  temp.cov,
  locations,
  B = 300,
  H0 = "ED",
  ms_models = c("powexp", "gauss", "brown"),
  sel_crit = "TIC",
  varmeth = "chain",
  maxiter = 300
)
}
\arguments{
\item{data}{A matrix of observations. Each column corresponds to one station.}

\item{temp.cov}{A numeric vector containing values of the temporal covariate. Must be of length \code{nrow(data)}.}

\item{locations}{A matrix containing the coordinates (longitude and latitude) of the stations.
The i-th row corresponds to the data contained in the i-th column of the data matrix.}

\item{B}{The number of bootstrap replicates.}

\item{H0}{Character code for the null-hypothesis: \code{ED} for the one of equal distribution;
\code{LS} for the one of a local scaling model.}

\item{ms_models}{A vector containing the names of the max-stable models to fit. Must be a subset of
\verb{gauss, brown, powexp, cauchy, whitmat, bessel}.}

\item{sel_crit}{The selection criterion based on which the max-stable model is selected. Must be one of
\verb{TIC, AIC}.}

\item{varmeth}{Method for estimating the variance-covariance matrix of the stationwise ML estimators, passed to
\code{\link[findpoolreg]{fit_spat_scalegev}}.
Can be either \code{chain} (the default) for an estimator based
on the multivariate chain rule, or \code{basic} for a very simplistic but faster method.}

\item{maxiter}{Passed to the optimisation function, giving the maximum number
of iterations used for optimising the Likelihood.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{teststat} The value of the test statistic on the observed data
\item \code{p} The asymptotic p value of the test statistic
\item \code{p_boot} The bootstrapped p-value.
\item \code{bootstrap.B} The number of bootstrap repetitions that were performed.
\item \code{Model_Sel} The max-stable process model that was selected.
\item \code{ms_pars} Parameters of the selected max-stable process model.
\item \code{boot_teststat} Data frame containing the test statistic and corresponding p-value for each bootstrap replicate.
\item \code{par_h0} The parameter estimates under H0.
\item \code{H0} Character code of the null hypothesis.
}
}
\description{
Bootstrap p-values for test-statistic
}
\examples{
\dontrun{
# generate some data
cvrt <- (1:100)/100
coords <- matrix(20*abs(stats::rnorm(4*2)), ncol = 2)
x <- generateData(seed = 2, n = 100, temp.cov = cvrt, d= 4, locations = coords)

bootres <- bootstrap_scalegev(data = x, temp.cov = cvrt, locations = coords,
                              varmeth = "chain", B = 200)
bootres
}
}
