% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_fun.R
\name{bootstrap_subsets_ms}
\alias{bootstrap_subsets_ms}
\title{Bootstrap p-values of test-statistic on several subsets}
\usage{
bootstrap_subsets_ms(
  data,
  temp.cov,
  locations,
  B = 300,
  H0 = "ED",
  ms_models = c("powexp", "gauss", "brown"),
  sel_crit = "TIC",
  varmeth = "chain",
  subsets,
  return_boots = FALSE,
  adj_pvals = FALSE,
  adj_method = "BH",
  set_start_vals = FALSE,
  maxiter = 300
)
}
\arguments{
\item{data}{A matrix of observations. Each column corresponds to one station.
Columns should be named or numbered to identify subsets.}

\item{temp.cov}{A numeric vector containing values of the temporal covariate. Must be of length \code{nrow(data)}.}

\item{locations}{A matrix containing the coordinates (longitude and latitude) of the stations.
The i-th row corresponds to the data contained in the i-th column of the data matrix.}

\item{B}{The number of bootstrap replicates.}

\item{H0}{Character code for the null-hypothesis: \code{ED} for the one of equal distribution;
\code{LS} for the one of a local scaling model.}

\item{ms_models}{A vector containing the names of the max-stable models to fit. Must be a subset of
\verb{gauss, brown, powexp, cauchy, whitmat, bessel}.}

\item{sel_crit}{The selection criterion based on which the max-stable model is selected. Must be one of
\verb{TIC, AIC}.}

\item{varmeth}{Method for estimating the variance-covariance matrix of the stationwise ML estimators, passed to
\code{\link[findpoolreg]{fit_spat_scalegev}}.
Can be either \code{chain} (the default) for an estimator based
on the multivariate chain rule, or \code{basic} for a very simplistic but faster method.}

\item{subsets}{A list containing subsets of columns on which to perform the bootstrap test.}

\item{return_boots}{Logical; whether to return bootstrapped values of test statistic and p-value.}

\item{adj_pvals}{Logical; whether to return column containing adjusted p-values.}

\item{adj_method}{Method used for adjusting the p-values. Passed to \code{\link[stats]{p.adjust}}.}

\item{set_start_vals}{Logical; whether to set the start values for the optimisation carried out on the bootstrap
to the estimates obtained under H0 on input data}

\item{maxiter}{Passed to the optimisation function, giving the maximum number
of iterations used for optimising the Likelihood.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{teststat} The value of the test statistic on the observed data
\item \code{p} The asymptotic p value of the test statistic
\item \code{p_boot} The bootstrapped p-value.
\item \code{bootstrap.B} The number of bootstrap repetitions that were performed.
\item \code{Model_Sel} The max-stable process model that was selected.
\item \code{ms_pars} Parameters of the selected max-stable process model.
\item \code{boot_teststat} (if \code{return_boots = TRUE}) Data frame containing the test statistic and corresponding p-value for each bootstrap replicate.
\item \code{par_h0} The parameter estimates under H0.
\item \code{H0} Character code of the null hypothesis.
\item \code{adj_p} (if \code{adj_pvals = TRUE}) The Holm-adjusted p-values.
}

Further, we have the columns
\itemize{
\item \code{X1} and \code{X2} when testing pairs, giving the labels of the respective pair
\item \code{sbst} when testing triples or larger subsets, containing a list of the respective subset.
}
}
\description{
Bootstrap p-values of test-statistic on several subsets
}
\examples{

\dontrun{
# generate 6-dimensional data where two stations deviate from the rest
cvrt <- (1:100)/100
coords <- matrix(20*abs(stats::rnorm(6*2)), ncol = 2)
x <- generateData(seed = 2, n = 100, temp.cov = cvrt, d= 6, locations = coords,
                 loc = c(rep(10,4), c(8, 8)), scale = c(rep(2,4), c(1.5, 1.5)),
                 shape = rep(0.1, 6), alpha = rep(2,6))

sbsts <- list( c(1,2), c(1,3), c(1,4), c(1,5), c(1,6))
bootres <- bootstrap_subsets_ms(data = x, temp.cov = cvrt,
             locations = coords, varmeth = "chain", B = 200, subsets = sbsts,
             adj_pvals = TRUE)
bootres # H0 ist rejected at 5\%-level for the deviating stations 5 and 6
}
}
