% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rl.R
\name{compute_rl}
\alias{compute_rl}
\title{Compute RL of GEV distribution in reference climate}
\usage{
compute_rl(theta, Tyrl, type, ref_gmst = NULL, ...)
}
\arguments{
\item{theta}{Estimated GEV parameters}

\item{Tyrl}{The return periods for which the RLs are computed}

\item{type}{The model: 'stationary'. 'shift' or 'scale'}

\item{ref_gmst}{The reference value of the temporal covariate for which the
RL is computed. Ignored when \code{type = "stationary"}.}
}
\value{
A tibble containg the estimated RL for each combination of \code{Tyrl} and
\code{ref_gmst}.
}
\description{
Compute RL of GEV distribution in reference climate
}
\examples{
compute_rl(theta = c(mu = 20, sigma = 2, gamma = 0.2, alpha = 2),
Tyrl = c(50, 100), type = "scale", ref_gmst = c(0.5, 0.95))

}
