% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststat_fun.R
\name{compute_teststat}
\alias{compute_teststat}
\title{Compute a data frame containing the value of the test statistic and p-value from the data}
\usage{
compute_teststat(data, temp.cov, equal_distr = TRUE)
}
\arguments{
\item{data}{Matrix or data frame of observations. Each column corresponds to one station/region.}

\item{temp.cov}{Temporal covariate with same record length as data.}

\item{equal_distr}{Logical; if \code{TRUE} (the default) the hypothesis is the one of equal distributions across
the different stations. If \code{FALSE}, the hypothesis is that the local scaling model holds across the region.}
}
\value{
A data frame with columns
\itemize{
\item teststat: observed value of the test statistic
\item p: p-value obtained from asymptotic \eqn{\chi^2} distribution (not reliable)
}
}
\description{
Compute a data frame containing the value of the test statistic and p-value from the data
}
\examples{
 # generate 3 dimensional data:
xx <- matrix(rep(exp((1:100)/100), 3)*evd::rgev(300), ncol = 3)
compute_teststat(data = xx, temp.cov = (1:100)/100, equal_distr = TRUE)
}
