% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cov.R
\name{estimate_gammas}
\alias{estimate_gammas}
\title{Computes cross-covariance matrix of score function applied to columns of
data transformed to standard GEV distribution}
\usage{
estimate_gammas(data, parmat, temp.cov = NULL, rel_par = FALSE)
}
\arguments{
\item{data}{Matrix of observations.}

\item{parmat}{Matrix of parameters of the scale-GEV-model. Each column corresponds to one station.}

\item{temp.cov}{Values of temporal covariate; numeric vector of length as \code{nrow(data)}.}

\item{rel_par}{Logical; whether \code{theta} contains the plain parameters
\eqn{\mu, \sigma, \gamma, \alpha},
or relative values \eqn{\mu, \mu/\sigma, \gamma, \alpha/\mu}.}
}
\value{
A dxd matrix.
}
\description{
Computes cross-covariance matrix of score function applied to columns of
data transformed to standard GEV distribution
}
\examples{
# generate data
cvrt <- (1:50)/50
d <- 4
coords <- matrix(20*abs(stats::rnorm(d*2)), ncol = 2)
x <- generateData(seed = 1, n = 50, temp.cov = cvrt, d= d, locations = coords)
mle <- fit_spat_scalegev(x, temp.cov =cvrt)
estimate_gammas(x, parmat = mle$mle, temp.cov =cvrt, rel_par = FALSE)
}
