% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data set}
\format{
A 75 x 9 matrix of observations sampled from a scale-GEV model with spatial dependence structure
described by a max-stable process.
}
\usage{
example_data
}
\description{
Example data set
}
\examples{
The data was generated with the following code.
meancoords <-  as.matrix(findpoolreg::example_grid[ , 6:7])
cvrt <- (-20:54)/100
generateData(n = 75, d = 9,
            scale = c(rep(6, 3), rep(5, 3), rep(3, 3)),
            loc =  c(rep(24, 3), rep(20, 3), rep(17, 3)),
            shape =  c(rep(0.1, 6), rep(0.15, 3)),
            alpha =  c(rep(3, 3), rep(2.5, 3),  rep(2, 3)),
            seed = 1, locations = meancoords, temp.cov = cvrt)

}
\keyword{datasets}
