% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_grid.R
\docType{data}
\name{example_grid}
\alias{example_grid}
\title{Example of grid coordinates}
\format{
A tibble with 7 columns giving  a region label, the edges of the grid tiles
and the centre points of the grid tiles.
}
\usage{
example_grid
}
\description{
Example of grid coordinates
}
\keyword{datasets}
