% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpoolreg-package.R
\docType{package}
\name{findpoolreg-package}
\alias{findpoolreg}
\alias{findpoolreg-package}
\title{findpoolreg: Find Pooling Region for Precipitation Extremes}
\description{
Parametric bootstrap test for assessing whether precipitation extremes of a set of locations are homogeneous w.r.t. their Generalised Extreme Value distribution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leandrazan/findpoolreg}
  \item Report bugs at \url{https://github.com/leandrazan/findpoolreg/issues}
}

}
\keyword{internal}
