% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_regionwise_rl.R
\name{gen_gev_ms_sample}
\alias{gen_gev_ms_sample}
\title{Generate one observation with a given dependence structure and given GEV
parameters}
\usage{
gen_gev_ms_sample(ms_fit, gev_par, coords, ref_gmst, seed)
}
\arguments{
\item{ms_fit}{List containing the max-stable process parameters, i.e.
\itemize{
\item \code{covmod}: A character string giving the covariance model of the max-stable process model
(one of \verb{gauss, brown, powexp, cauchy, whitmat, bessel}.)
\item \code{parhat}: The parameters of the respective covariance model.
}}

\item{gev_par}{Values of the scale-GEV parameters, in the order \eqn{\mu, \sigma,\gamma, \alpha}.}

\item{coords}{The coordinates (longitude and latitude) of the locations for which you
want to generate data.}

\item{ref_gmst}{The reference value of the temporal covariate for which data is generated.}

\item{seed}{A random seed.}
}
\value{
A vector of length \code{nrow(coords)} containing one observation for each location.
}
\description{
Generate one observation with a given dependence structure and given GEV
parameters
}
\examples{

d <- 10
coords <- matrix(20*abs(stats::rnorm(d*2)), ncol = 2)
gen_gev_ms_sample(ms_fit = list(covmod = "gauss", parhat = c(1.2, 0.1, 1.2)),
gev_par = c(20, 4, 0.2, 2), coords = coords, ref_gmst = 0.9, seed = 1)

}
