% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_regionwise_rl.R
\name{gen_sample_max}
\alias{gen_sample_max}
\title{Generate sample of independent regionwise maxima}
\usage{
gen_sample_max(ms_fit, gev_par, ref_gmst, B, coords, start = 0)
}
\arguments{
\item{ms_fit}{List containing the max-stable process parameters, i.e.
\itemize{
\item \code{covmod}: A character string giving the covariance model of the max-stable process model
(one of \verb{gauss, brown, powexp, cauchy, whitmat, bessel}.)
\item \code{parhat}: The parameters of the respective covariance model.
}}

\item{gev_par}{Values of the scale-GEV parameters, in the order \eqn{\mu, \sigma,\gamma, \alpha}.}

\item{ref_gmst}{The reference value of the temporal covariate for which data is generated.}

\item{B}{The resulting sample length.}

\item{coords}{The coordinates (longitude and latitude) of the locations for which you
want to generate data.}

\item{start}{Start value for the random seed; \code{gen_gev_ms_sample} is called with
seeds  \code{(start + 1) : (start + B)}.}
}
\value{
A vector of length \code{B} containing simulated regionwise maxima.
}
\description{
Calls the function \code{gen_gev_ms_sample} repeatedly and computes maximum of each
iteration.
}
\examples{

d <- 10
coords <- matrix(20*abs(stats::rnorm(d*2)), ncol = 2)
maxis <- gen_sample_max(ms_fit = list(covmod = "gauss", parhat = c(1.2, 0.1, 1.2)),
   gev_par = c(20, 4, 0.2, 2), ref_gmst = 0.9, B = 100, coords = coords)
plot.ts(maxis)
}
