% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_fun.R
\name{generate_bootsamp_unitfrech}
\alias{generate_bootsamp_unitfrech}
\title{Generate bootstrap samples with dependence structure from data and unit frechet margins}
\usage{
generate_bootsamp_unitfrech(
  data,
  temp.cov,
  locations,
  ms_models = c("powexp", "gauss", "brown"),
  B = 500,
  sel_crit = "TIC",
  warn_msfit = TRUE
)
}
\arguments{
\item{data}{Numeric matrix of observations. Each column corresponds to one station.}

\item{temp.cov}{Numeric vector of temporal covariate, with same length as \code{nrow(data)}}

\item{locations}{A matrix containing longitude and latitude of the stations. Each row corresponds to one station.}

\item{ms_models}{A vector containing the names of the max-stable models to fit. Must be a subset of
\verb{gauss, brown, powexp, cauchy, whitmat, bessel}.}

\item{B}{The number of bootstrap samples that are generated.}

\item{sel_crit}{The selection criterion based on which the max-stable model is selected. Must be one of
\verb{TIC, AIC}.}

\item{warn_msfit}{logical; whether to print warnings possibly generated during fitting the max-stable process.}
}
\value{
Returns a list of  \code{B} bootstrap samples, each being a matrix of same dimension as input data.
}
\description{
This function transform margins to unit Frechet, fits several max-stable models to
the transformed data and chooses the best fit. Then ootstrap samples from this fitted max-stable model
are generated.
}
\examples{
# generate data

d <- 8
coords <-  matrix(20*abs(stats::rnorm(d*2)), ncol = 2)
xx <- generateData(seed = 1, d = d, locations = coords,
 scale = c(1:d), loc = rep(20, d), shape = seq(0, 0.3, length.out = d), alpha = rep(2.5, d))

# generate bootstrap samples
bootsamps <- generate_bootsamp_unitfrech(xx, temp.cov = (1:100)/100, locations = coords,
                                        ms_models = c("powexp", "gauss", "brown"), B = 2,
                                         sel_crit = "TIC", warn_msfit = FALSE)
plot.ts(bootsamps[[1]])
plot.ts(bootsamps[[2]])
}
