% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generateData}
\alias{generateData}
\title{Sample from a spatial scale-GEV-model}
\usage{
generateData(
  n = 100,
  d = 9,
  scale = rep(5, d),
  loc = rep(20, d),
  shape = rep(0.1, d),
  alpha = rep(2.5, d),
  covmod = "gauss",
  locations = matrix(20 * abs(stats::rnorm(d * 2)), ncol = 2),
  temp.cov = (1:100)/100,
  MSparam = list(cov11 = 0.4, cov12 = 0.2, cov22 = 0.9, range = 0.55, smooth = 1),
  seed
)
}
\arguments{
\item{n}{Record length}

\item{d}{Number of stations (i.e. the dimension of the sampled data)}

\item{scale}{vector of length \eqn{d} containing the scale parameters for each station}

\item{loc}{vector of length \eqn{d} containing the location parameters for each station}

\item{shape}{vector of length \eqn{d} containing the shape parameters for each station}

\item{alpha}{vector of length \eqn{d} containing the trend parameters for each station}

\item{covmod}{The covariance model for the max-stable-process. Must be one of \code{gauss},
\code{brown},  \code{powexp}, \code{cauchy}, \code{whitmat}, \code{bessel}.}

\item{locations}{A \eqn{d \times 2} matrix containing longitude and latitude of the \eqn{d} stations.}

\item{temp.cov}{A vector of length \eqn{n} giving the temporal covariate.}

\item{MSparam}{A named list containing the parameters for the max-stable process. For \code{gauss}, these must
be \verb{cov11, cov12, cov22}, for the other models these are \verb{range, smooth}. For further details,
see \link[SpatialExtremes]{rmaxstab}.}

\item{seed}{This seed is set for reproducibility.}
}
\value{
A \eqn{n \times d} matrix containing containing the data.
}
\description{
This function generates data with a max-stable-dependence structure and margins from
the scale-GEV-model.
}
\examples{

# with the same scale-GEV model parameters for d stations
plot.ts(generateData(seed = 1))
# with different scale-GEV model parameters at station 1 and 2
plot.ts(generateData(scale = c(2, 2, rep(5, 7)), shape =  c(0.2, 0.2, rep(0.1, 7)),  seed = 1))
}
