% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_regionwise_rl.R
\name{get_gev_par}
\alias{get_gev_par}
\title{Evaluate parameters of scale-GEV at covariate values}
\usage{
get_gev_par(parvec, cvrt)
}
\arguments{
\item{parvec}{Vector of scale-GEV parameters, in the order \eqn{\mu, \sigma,\gamma, \alpha}.}

\item{cvrt}{Vector of reference covariate values.}
}
\value{
A vector of length 3 (when \code{cvrt} consists of only one reference value) or a
matrix of dimension 3 x \code{length(cvrt)}, containing values of
\eqn{\mu(t), \sigma(t), \gamma}.
}
\description{
Evaluate parameters of scale-GEV at covariate values
}
\examples{
get_gev_par(c(20, 4, 00.2, 2), cvrt = c(0.9, 0.99))
}
