% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_regionwise_rl.R
\name{get_regionwise_estimates}
\alias{get_regionwise_estimates}
\title{Estimate 'regionwise' RP and RL for pooled region}
\usage{
get_regionwise_estimates(
  data,
  coords,
  temp.cov,
  ref_gmst,
  B,
  r = NULL,
  T.year = NULL,
  ms_models = c("gauss", "brown", "whitmat", "powexp"),
  return_ecdf = TRUE,
  start = 0
)
}
\arguments{
\item{data}{Data matrix containing data of the pooled region.}

\item{coords}{Coordinates of the locations that make up the pooled region.}

\item{temp.cov}{Numeric vector of same length \code{nrow(data)} containing the
temporal covariate.}

\item{ref_gmst}{Reference value of temporal covariate for which to compute the
regionwise estimates.}

\item{B}{A large integer giving the sample size based on which the RP/RL is
estimated.}

\item{r}{A numeric vector of values for which to compute the regionwise return
period.}

\item{T.year}{A numeric vèctor of values for which to compute the regionwise
return level.}

\item{ms_models}{A list of max-stable process models that are fitted to the
data in order to find the best fit.}

\item{return_ecdf}{Logical; whether to return the empirical cumulative distribution
function of the simulated samples.}

\item{start}{Passed to \code{\link[findpoolreg]{gen_sample_max}}, giving start
value for random seeds.}
}
\value{
A list containing a subset of the following components:
\itemize{
\item \code{ecdf_max} The empirical cumulative distribution function of the simulated
sample of regionwise maxima.
\item \code{rl_reg} The estimated regionwise RLs.
\item \code{rp_reg} The estimated regionwise RPs.
\item \code{gev_par} The estimated scale-GEV parameters of the pooled sample.
\item \code{fit_ms} A list containing the components \code{covmod} and \code{parhat}, giving
the (covariance) model and the respective estimated parameters.
}
}
\description{
Estimate 'regionwise' RP and RL for pooled region
}
\examples{
d <- 10
coords <- matrix(20*abs(stats::rnorm(d*2)), ncol = 2)

 xx <- generateData(locations = coords, d = 10, seed = 2)
get_regionwise_estimates(data = xx, coords = coords, temp.cov = (1:100)/100,
    ref_gmst = 0.9, B = 10000, T.year = 100, r = 65)
}
