% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_fun.R
\name{grad_ll_scalegev}
\alias{grad_ll_scalegev}
\title{Gradient of log-density in scale-GEV-model}
\usage{
grad_ll_scalegev(data, params, temp.cov)
}
\arguments{
\item{data}{Numeric vector of observations}

\item{params}{Vector of scale-GEV-model parameters (length 4)}

\item{temp.cov}{Values of temporal covariate (with same length as data)}
}
\value{
A \eqn{ 4\times} \code{length(data)}-matrix.
}
\description{
Gradient of log-density in scale-GEV-model
}
\examples{
xx <- exp((1:100)/100)*evd::rgev(100)
grad_ll_scalegev(xx, params = c(1, 1, 0, 1), temp.cov = (1:100)/100)
}
