% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_fun.R
\name{nll_scalegev_hom}
\alias{nll_scalegev_hom}
\title{Negative log-Likelihood of spatial scale-GEV-model under homogeneity constraint}
\usage{
nll_scalegev_hom(params, data, temp.cov)
}
\arguments{
\item{params}{Vector of length \eqn{d +3}, where \eqn{d} is the number of sites.
Parameters must be given in the order
\eqn{(\mu_1, \mu_2, \ldots, \mu_d, \delta, \gamma, \eta)}.}

\item{data}{Matrix or data frame containing the observations from \eqn{d} measuring sites
(one column per site)}

\item{temp.cov}{Vector containing values of the temporal covariate.}
}
\value{
Numeric value of the log-Likelihood of the data evaluated at the given parameters.
}
\description{
Negative log-Likelihood of spatial scale-GEV-model under homogeneity constraint
}
\examples{
# generate some data
xx <- (1:3)*matrix(rep(exp((1:100)/100), 3)*evd::rgev(300), ncol = 3)

nll_scalegev_hom(params = c(1, 1, 1, 1, 0, 1), data = xx, temp.cov = (1:100)/100)
}
