% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_pvalues.R
\name{plot_regions}
\alias{plot_regions}
\title{Visualise bootrstrap test results for pairs}
\usage{
plot_regions(
  coord_grid,
  testres,
  rej_proc = "Holm",
  plot_type = "rejection",
  level = 0.1,
  bins = c(0, 0.005, 0.01, 0.02, 0.05, 1),
  ...
)
}
\arguments{
\item{coord_grid}{A dataframe containing the grid coordinates in the columns \verb{from_lon,to_lon, from_lat, to_lat}.
Can also contain the grid centers through columns \code{meanlon}, \code{meanlat}.}

\item{testres}{Result of bootstrap test as obtained from  \code{\link[findpoolreg]{bootstrap_scalegev_subsets}}.}

\item{rej_proc}{The rejection procedure: either \code{Holm} for the Holm stepdown procedure, or \code{BenjYek} for the
Benjamini Yekutieli stepup procedure. Only relevant when \code{plot_type = rejection}.}

\item{plot_type}{The plot type: either \code{rejection}, then regions are shaded according to test result,
i.e. rejection or no rejection or \code{pvals}, then  regions are shaded according to their p-value.}

\item{level}{The test level}

\item{bins}{Bins for p-value shades when \code{plot_type = pvals}.}

\item{...}{additional arguments passed to \code{\link[leaflet]{leafletOptions}}}
}
\value{
A leaflet plot.
}
\description{
Visualise bootrstrap test results for pairs
}
\examples{

## generate data and compute test results
\dontrun{
coord_grid <- tidyr::expand_grid(from_lat = seq(44.75, 48.5, 1.25), from_lon = seq(7.25, 13.25, 2) ) \%>\%
         dplyr::mutate(to_lon = from_lon + 2, to_lat = from_lat + 1.25, Region = 1:16)

meancoords <-   as.matrix((coord_grid \%>\% dplyr::mutate(meanlon = (from_lon + to_lon)/2,
                                                           meanlat = (from_lat + to_lat)/2))[ , 6:7])
cvrt <- (-20:79)/100
simdat <- generateData(n = 100, d = 16, scale = c(rep(6, 3), rep(5, 10), rep(3, 3)),
                       loc =  c(rep(24, 3), rep(20, 7), rep(17, 6)), shape =  c(rep(0.1, 13), rep(0.15, 3)),
                       alpha =  c(rep(3, 3), rep(2.5, 7),  rep(2, 6)),
seed = 1, locations = meancoords, temp.cov =  cvrt)

subsets <- purrr::map(c(1:9, 11:16), ~ c(10, .x))
bootres <- bootstrap_scalegev_subsets(data = simdat, temp.cov = cvrt, locations = meancoords, subsets = subsets)

plot_regions(coord_grid = coord_grid, testres = bootres, rej_proc = "Holm", zoom)
plot_regions(coord_grid = coord_grid, testres = bootres, rej_proc = "BenjYek")

plot_regions(coord_grid = coord_grid, testres = bootres, plot_type = "pvals")
}
}
