% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalegev2frech_fun.R
\name{scalegev2frech}
\alias{scalegev2frech}
\title{Transform scale-GEV margins to unit Frechet margins}
\usage{
scalegev2frech(data, temp.cov, par = NULL, maxiter = 300)
}
\arguments{
\item{data}{Numeric vector of observations}

\item{temp.cov}{Numeric vector of temporal covariate, with same length as \code{data}}

\item{par}{Either \code{NULL} (the default) or a vector containing the location, scale, shape and
trend parameter of the scale-GEV model. If \code{NULL}, the parameters are estimated with
maximum likelihood.}

\item{maxiter}{Only relevant when \code{par = NULL}; passed to the optimisation
function, giving the maximum number of iterations used for optimising.}
}
\value{
A vector with same length as input data.
}
\description{
Transform scale-GEV margins to unit Frechet margins
}
\examples{
xx <- generateData(n = 100, temp.cov = (1:100)/100, seed=1)

# apply transformation on first column
xx[, 1] <- scalegev2frech(xx[, 1], temp.cov = (1:100)/100)

# spot the difference
plot.ts(xx)
}
