% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cov.R
\name{score_standard_univ}
\alias{score_standard_univ}
\title{Computes three dimensional score function of data with margins transformed to standard GEV}
\usage{
score_standard_univ(x, theta, temp.cov = NULL, rel_par = FALSE)
}
\arguments{
\item{x}{Numeric vector of observations from the scale-GEV-model}

\item{theta}{Parameter vector of the scale-GEV-model}

\item{temp.cov}{Values of temporal covariate; numeric vector of same length as x}

\item{rel_par}{Logical; whether \code{theta} contains the plain parameters \eqn{\mu, \sigma, \gamma, \alpha},
or relative values \eqn{\mu, \mu/\sigma, \gamma, \alpha/\mu}.}
}
\value{
A matrix of dimension \eqn{3\times} \code{length(x)} containing the values of
\deqn{ \dot\ell_{(0,1, \gamma)} \left( \frac{ x - \mu(gmst(t))}{\sigma(gmst(t))} \right).}
}
\description{
Computes three dimensional score function of data with margins transformed to standard GEV
}
\examples{
 xx <- exp((1:100)/100)*evd::rgev(100)
 score_standard_univ(x = xx, theta = c(10, 2, 0, 1), temp.cov = (1:100)/100, rel_par = FALSE)
}
