% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalegev2frech_fun.R
\name{spat_scalegev2frech}
\alias{spat_scalegev2frech}
\title{Transform scale-GEV margins to unit Frechet margins for multivariate data}
\usage{
spat_scalegev2frech(data, temp.cov, par = NULL, maxiter = 300)
}
\arguments{
\item{data}{Numeric matrix of observations. Each column corresponds to one station.}

\item{temp.cov}{Numeric vector of temporal covariate, with same length as \code{nrow(data)}}

\item{par}{Either \code{NULL} (the default) or a \eqn{4 \times ncol(data)} matrix containing the location, scale, shape and
trend parameters of the scale-GEV model of each column of the data. If \code{NULL}, the parameters are estimated stationwise with
maximum likelihood.}

\item{maxiter}{Only relevant when \code{par = NULL}; passed to the optimisation
function, giving the maximum number of iterations used for optimising.}
}
\value{
A matrix of same dimension as input data.
}
\description{
Transform scale-GEV margins to unit Frechet margins for multivariate data
}
\examples{
xx <- generateData(n = 100, temp.cov = (1:100)/100, seed=1)

# apply transformation
yy <- spat_scalegev2frech(xx, temp.cov = (1:100)/100)

# spot the difference
plot.ts(xx)
plot.ts(yy)
}
