% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_pvalues.R
\name{visualise_test_res}
\alias{visualise_test_res}
\title{Visualise bootstrap test results for pairwise testing}
\usage{
visualise_test_res(
  coord_grid,
  testres,
  method = "holm",
  plot_type = "rejection",
  level = 0.1,
  bins = c(0, 0.05, 0.075, 0.1, 1),
  loi = NULL,
  stationData = FALSE,
  ...
)
}
\arguments{
\item{coord_grid}{A dataframe containing the coordinates of the grid edges in columns
named \verb{from_lon, to_lon, from_lat, to_lat}.
Can also contain the grid centers named \code{meanlon}, \code{meanlat}. Further, the same region labels
used in the bootstrap procedure must be given in a column named \code{Region}.
When station data is used instead of
grid data, provide the coordinates of the stations in columns named \code{lon} and \code{lat} instead,
and the Station labels in a column named \code{Station}. In this case,
set \code{stationData = TRUE}.}

\item{testres}{Result of bootstrap test as obtained from  \code{\link[findpoolreg]{bootstrap_subsets_ms}} or
\code{\link[findpoolreg]{bootstrap_pairs_biv}}.}

\item{method}{The method used for adjusting p-values. Must be one of \code{holm} (for the Holm stepdown procedure),
\code{BY} (for the Benjamini Yekutieli stepup procedure), \code{BH} (for the Benjamini Hochberg step-up procedure) or
\code{boot} when not adjusting for multiple testing.}

\item{plot_type}{The plot type: either \code{rejection}, then regions are shaded according to test result,
i.e. rejection or no rejection or \code{pvals}, then  regions are shaded according to their (adjusted) p-values.}

\item{level}{The test level}

\item{bins}{Bins for p-value shades when \code{plot_type = pvals}.}

\item{loi}{Location(s) of interest, passed as numeric vector containing location labels.
Can be omitted when location of interest is of length 1.}

\item{stationData}{Logical; whether station data is used instead of gridded data.}

\item{...}{additional arguments passed to \code{leafletOptions} from \code{\link[leaflet]{leaflet}},
e.g. \code{zoomControl = FALSE}. Further,
the arguments \code{position} (one of \verb{bottomright, bottomleft, topright, topleft}) which gives the legend position,
\code{width} and \code{height} for the plot size can be given.}
}
\value{
A leaflet plot.
}
\description{
Visualise bootstrap test results for pairwise testing
}
\examples{

\dontrun{
## generate data and compute test results
coord_grid <- tidyr::expand_grid(from_lat = seq(44.75, 47.25, 1.25),
                                 from_lon = seq(7.25, 11.25, 2)) \%>\%
         dplyr::mutate(to_lon = from_lon + 2, to_lat = from_lat + 1.25, Region = 1:9)

meancoords <- as.matrix((coord_grid \%>\%
                   dplyr::mutate(meanlon = (from_lon + to_lon)/2,
                    meanlat = (from_lat + to_lat)/2))[ , 6:7])
cvrt <- (-20:79)/100
simdat <- generateData(n = 100, d = 9,
                       scale = c(rep(6, 3), rep(5, 3), rep(3, 3)),
                       loc =  c(rep(24, 3), rep(20, 3), rep(17, 3)),
                       shape =  c(rep(0.1, 6), rep(0.15, 3)),
                       alpha =  c(rep(3, 3), rep(2.5, 3),  rep(2, 3)),
                       seed = 1, locations = meancoords, temp.cov =  cvrt)

subsets <- purrr::map(c(1:4, 6:9), ~ c(5, .x))
bootres <- bootstrap_subsets_ms(data = simdat, temp.cov = cvrt, locations = meancoords,
 subsets = subsets, B = 100)

visualise_test_res(coord_grid = coord_grid, testres = bootres,
                   method = "holm", plot_type = "pvals", loi = 5)
visualise_test_res(coord_grid = coord_grid, testres = bootres, method = "BY",
                   loi = 5)

pvisualise_test_res(coord_grid = coord_grid, testres = bootres, method = "BH",
                    plot_type = "pvals", loi = 5)
}
}
