% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_fun.R
\name{adjust_pvals}
\alias{adjust_pvals}
\title{Holms-correction of p-values for multiple testing}
\usage{
adjust_pvals(pvaltib, nhyp)
}
\arguments{
\item{pvaltib}{Tibble}

\item{nhyp}{}
}
\value{

}
\description{
Holms-correction of p-values for multiple testing
}
