% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststat_fun.R
\name{compute_teststat}
\alias{compute_teststat}
\title{Compute a data frame containing the value of the test statistic and p-value from the data}
\usage{
compute_teststat(
  data,
  temp.cov,
  H0 = "ED",
  varmeth = "chain",
  start_vals = NULL
)
}
\arguments{
\item{data}{Matrix or data frame of observations. Each column corresponds to one station/region.}

\item{temp.cov}{Temporal covariate with same record length as data.}

\item{H0}{Either 'ED' for 'Equal distribution, i.e. the hypothesis is the one of equal distributions across
the different stations, or 'LS' for 'Local Scaling', i.e. the hypothesis is that the local scaling model holds across the region.}

\item{varmeth}{Method for estimation of variance-covariance matrix. Can be either \code{chain} (the default) for an estimator based
on the multivariate chain rule, or \code{basic} for a very simplistic but faster method. Passed to
\link[findpoolreg]{fit_spat_scalegev}.}

\item{start_vals}{Optional: matrix containing start values for the ML optimisation.}
}
\value{
A data frame with columns
\itemize{
\item teststat: observed value of the test statistic
\item p: p-value obtained from asymptotic \eqn{\chi^2} distribution (not reliable)
}
}
\description{
Compute a data frame containing the value of the test statistic and p-value from the data
}
\examples{
 # generate 3 dimensional data:
xx <- matrix(rep(exp((1:100)/100), 3)*evd::rgev(300), ncol = 3)
compute_teststat(data = xx, temp.cov = (1:100)/100, H0 = "ED")
}
