% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_fun.R
\name{fit_spat_scalegev}
\alias{fit_spat_scalegev}
\title{Fit spatial scale-GEV-model to data}
\usage{
fit_spat_scalegev(
  data,
  temp.cov,
  method = "BFGS",
  maxiter = 300,
  varmeth = "chain",
  start_vals = NULL
)
}
\arguments{
\item{data}{A matrix or dataframe representing the data, where each column corresponds to one site.}

\item{temp.cov}{Values of the temporal covariate of the scale-GEV-model.
Must be of same record length as thr observations.}

\item{method}{Method passed to \code{\link[stats:optim]{stats::optim()}}.}

\item{maxiter}{Maximum number of iterations during maximisation (also passed to \code{\link[stats:optim]{stats::optim()}}).}

\item{varmeth}{Method for estimation of variance-covariance matrix. Can be either \code{chain} (the default) for an estimator based
on the multivariate chain rule, or \code{basic} for a very simplistic but faster method.}

\item{start_vals}{Optional: matrix containing start values for the ML optimisation.}
}
\value{
A list containing the estimated parameter values in \code{mle} and
an estimation of the variance-covariance matrix.
}
\description{
Fit spatial scale-GEV-model to data
}
\examples{
# generate some data
xx <- (1:3)*matrix(rep(exp((1:100)/100), 3)*evd::rgev(300), ncol = 3)

# fit componentwise scale-GEV-model
fit_spat_scalegev(data = xx, temp.cov = (1:100)/100, hom = FALSE)$mle


}
